starter-kits/2-reasoning-agents/requirements.txt
# Microsoft Foundry SDK (AI Projects)
azure-ai-projects>=1.0.0b1
azure-identity>=1.15.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests (for API integrations)
httpx>=0.27.0
aiohttp>=3.9.0

# Data handling
pydantic>=2.0.0

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: For local development with Microsoft Agent Framework
# Uncomment the line below if using the Agent Framework locally
# agent-framework>=0.1.0
